Mewtwo
Love him
